+++ app/src/main/java/it/cammino/risuscito/CantiParolaFragment.kt
@@ -1,15 +1,9 @@
 package it.cammino.risuscito
 
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
 import android.content.Intent
 import android.os.Bundle
 import android.os.SystemClock
 import android.preference.PreferenceManager
-import com.google.android.material.snackbar.Snackbar
-import androidx.fragment.app.Fragment
-import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.LinearLayoutManager
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
@@ -18,9 +12,15 @@ import android.view.View.OnLongClickListener
 import android.view.ViewGroup
 import android.widget.TextView
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.isVisible
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.afollestad.materialcab.MaterialCab
 import com.crashlytics.android.Crashlytics
+import com.google.android.material.snackbar.Snackbar
 import com.mikepenz.community_material_typeface_library.CommunityMaterial
 import com.mikepenz.fastadapter.commons.adapters.FastItemAdapter
 import com.mikepenz.iconics.IconicsDrawable
@@ -55,7 +55,7 @@ class CantiParolaFragment : Fragment() {
     private var longclickedPos: Int = 0
     private var longClickedChild: Int = 0
     private var cantoAdapter: FastItemAdapter<ListaPersonalizzataItem>? = null
-    private var actionModeOk: Boolean = false
+    //    private var actionModeOk: Boolean = false
     private var mMainActivity: MainActivity? = null
     private var mLastClickTime: Long = 0
     private var mLUtils: LUtils? = null
@@ -98,7 +98,7 @@ class CantiParolaFragment : Fragment() {
 
             result.append(getTitoloToSendFromPosition(3))
             result.append("\n")
-            val pref = PreferenceManager.getDefaultSharedPreferences(activity)
+            val pref = PreferenceManager.getDefaultSharedPreferences(context)
 
             if (pref.getBoolean(Utility.SHOW_PACE, false)) {
                 result.append(resources.getString(R.string.canto_pace).toUpperCase(l))
@@ -132,7 +132,7 @@ class CantiParolaFragment : Fragment() {
 //        if (parent.findViewById<View>(R.id.addCantoGenerico).visibility == View.VISIBLE) {
         if (parent.findViewById<View>(R.id.addCantoGenerico).isVisible) {
             if (mSwhitchMode) {
-                actionModeOk = true
+//                actionModeOk = true
                 MaterialCab.destroy()
                 Thread(
                         Runnable {
@@ -173,7 +173,7 @@ class CantiParolaFragment : Fragment() {
                 } else
                     openPagina(v)
             else {
-                actionModeOk = true
+//                actionModeOk = true
                 MaterialCab.destroy()
                 Thread(
                         Runnable {
@@ -355,55 +355,56 @@ class CantiParolaFragment : Fragment() {
 
     private fun scambioCanto(v: View, position: Int) {
         val idNew = Integer.valueOf((v.findViewById<View>(R.id.text_id_canto_card) as TextView).text.toString())
-        val timestampNew = (v.findViewById<View>(R.id.text_timestamp) as TextView).text.toString()
         if (idNew != idDaCanc || posizioneDaCanc != position) {
-
-            val positionToDelete = CustomList()
-            positionToDelete.id = 1
-            positionToDelete.position = position
-            positionToDelete.idCanto = idNew
             val mDao = RisuscitoDatabase.getInstance(context!!).customListDao()
-            mDao.deletePosition(positionToDelete)
+            if (mDao.checkExistsPosition(1, position, idDaCanc) > 0
+                    || mDao.checkExistsPosition(1, posizioneDaCanc, idNew) > 0) {
+                Snackbar.make(rootView!!, R.string.present_yet, Snackbar.LENGTH_SHORT).show()
+            } else {
+                val positionToDelete = mDao.getPositionSpecific(1, position, idNew)
+                mDao.deletePosition(positionToDelete)
 
-            mDao.updatePositionNoTimestamp(idNew, 1, posizioneDaCanc, idDaCanc)
+                mDao.updatePositionNoTimestamp(idNew, 1, posizioneDaCanc, idDaCanc)
 
-            val positionToInsert = CustomList()
-            positionToInsert.id = 1
-            positionToInsert.position = position
-            positionToInsert.idCanto = idDaCanc
-            positionToInsert.timestamp = Date(java.lang.Long.parseLong(timestampNew))
-            mDao.insertPosition(positionToInsert)
-            Snackbar.make(
-                    activity!!.findViewById<View>(R.id.main_content),
-                    R.string.switch_done,
-                    Snackbar.LENGTH_SHORT)
-                    .show()
+                val positionToInsert = CustomList()
+                positionToInsert.id = 1
+                positionToInsert.position = position
+                positionToInsert.idCanto = idDaCanc
+                positionToInsert.timestamp = positionToDelete.timestamp
+                mDao.insertPosition(positionToInsert)
+
+                Snackbar.make(
+                        activity!!.findViewById<View>(R.id.main_content),
+                        R.string.switch_done,
+                        Snackbar.LENGTH_SHORT)
+                        .show()
+            }
         } else {
             Snackbar.make(rootView!!, R.string.switch_impossible, Snackbar.LENGTH_SHORT).show()
         }
     }
 
     private fun scambioConVuoto(position: Int) {
-
-        val positionToDelete = CustomList()
-        positionToDelete.id = 1
-        positionToDelete.position = posizioneDaCanc
-        positionToDelete.idCanto = idDaCanc
         val mDao = RisuscitoDatabase.getInstance(context!!).customListDao()
-        mDao.deletePosition(positionToDelete)
+        if (mDao.checkExistsPosition(1, position, idDaCanc) > 0)
+            Snackbar.make(rootView!!, R.string.present_yet, Snackbar.LENGTH_SHORT).show()
+        else {
+            val positionToDelete = mDao.getPositionSpecific(1, posizioneDaCanc, idDaCanc)
+            mDao.deletePosition(positionToDelete)
 
-        val positionToInsert = CustomList()
-        positionToInsert.id = 1
-        positionToInsert.position = position
-        positionToInsert.idCanto = idDaCanc
-        positionToInsert.timestamp = Date(java.lang.Long.parseLong(timestampDaCanc!!))
-        mDao.insertPosition(positionToInsert)
+            val positionToInsert = CustomList()
+            positionToInsert.id = 1
+            positionToInsert.position = position
+            positionToInsert.idCanto = idDaCanc
+            positionToInsert.timestamp = positionToDelete.timestamp
+            mDao.insertPosition(positionToInsert)
 
-        Snackbar.make(
-                activity!!.findViewById<View>(R.id.main_content),
-                R.string.switch_done,
-                Snackbar.LENGTH_SHORT)
-                .show()
+            Snackbar.make(
+                    activity!!.findViewById<View>(R.id.main_content),
+                    R.string.switch_done,
+                    Snackbar.LENGTH_SHORT)
+                    .show()
+        }
     }
 
     private fun startCab(switchMode: Boolean) {
@@ -430,7 +431,7 @@ class CantiParolaFragment : Fragment() {
                         .sizeDp(24)
                         .paddingDp(2)
                         .colorRes(android.R.color.white)
-                actionModeOk = false
+//                actionModeOk = false
             }
 
             onSelection { item ->
@@ -447,7 +448,7 @@ class CantiParolaFragment : Fragment() {
                                     mDao.deletePosition(positionToDelete)
                                 })
                                 .start()
-                        actionModeOk = true
+//                        actionModeOk = true
                         MaterialCab.destroy()
                         Snackbar.make(
                                 activity!!.findViewById(R.id.main_content),
@@ -486,16 +487,16 @@ class CantiParolaFragment : Fragment() {
             }
 
             onDestroy {
-                Log.d(TAG, "MaterialCab onDestroy: $actionModeOk")
+                //                Log.d(TAG, "MaterialCab onDestroy: $actionModeOk")
                 mSwhitchMode = false
-                if (!actionModeOk) {
-                    try {
-                        posizioniList[longclickedPos].listItem!![longClickedChild].setmSelected(false)
-                        cantoAdapter!!.notifyItemChanged(longclickedPos)
-                    } catch (e: Exception) {
-                        Crashlytics.logException(e)
-                    }
+//                if (!actionModeOk) {
+                try {
+                    posizioniList[longclickedPos].listItem!![longClickedChild].setmSelected(false)
+                    cantoAdapter!!.notifyItemChanged(longclickedPos)
+                } catch (e: Exception) {
+                    Crashlytics.logException(e)
                 }
+//                }
                 true
             }
         }
@@ -524,7 +525,7 @@ class CantiParolaFragment : Fragment() {
                             posizioniList.add(
                                     getCantofromPosition(mCanti, getString(R.string.terza_lettura), 4, progressiveTag++))
 
-                            val pref = PreferenceManager.getDefaultSharedPreferences(activity)
+                            val pref = PreferenceManager.getDefaultSharedPreferences(context)
                             if (pref.getBoolean(Utility.SHOW_PACE, false))
                                 posizioniList.add(
                                         getCantofromPosition(mCanti, getString(R.string.canto_pace), 6, progressiveTag++))
